@@ -5,15 +5,15 @@ import 'package:animated_indexed_stack/stack_page.dart';
 
 class AnimatedIndexedStack extends StatefulWidget {
   @required
-  final int selectedIndex;
-  final List<Widget> children;
-  final List<RoutePageBuilder> pageBuilderList;
-  final Duration duration;
-  final RouteTransitionsBuilder transitionBuilder;
-  final SORT_TIME sortTime;
+  final int? selectedIndex;
+  final List<Widget>? children;
+  final List<RoutePageBuilder>? pageBuilderList;
+  final Duration? duration;
+  final RouteTransitionsBuilder? transitionBuilder;
+  final SORT_TIME? sortTime;
 
   const AnimatedIndexedStack({
-    Key key,
+    Key? key,
     this.pageBuilderList,
     this.children,
     this.selectedIndex,
@@ -29,11 +29,11 @@ class AnimatedIndexedStack extends StatefulWidget {
 class _AnimatedIndexedStackState extends State<AnimatedIndexedStack>
     with TickerProviderStateMixin {
   Map<int, Key> _pageIndexToKeyMap = new Map<int, Key>();
-  List<Widget> _pages;
-  List<dynamic> _list;
-  int _selectedIndex;
+  late List<Widget> _pages;
+  List<dynamic>? _list;
+  int? _selectedIndex;
 
-  int _prevPage = 1;
+  int? _prevPage = 1;
   Map<Key, AnimationController> _animationControllerList =
       new Map<Key, AnimationController>();
   Map<Key, AnimationController> _secondaryAnimationControllerList =
@@ -55,7 +55,7 @@ class _AnimatedIndexedStackState extends State<AnimatedIndexedStack>
     } else {
       _list = widget.pageBuilderList;
     }
-    _pages = _list.asMap().entries.map((entry) {
+    _pages = _list!.asMap().entries.map((entry) {
       final _index = entry.key;
 
       // Gerate Unique keys
@@ -111,19 +111,19 @@ class _AnimatedIndexedStackState extends State<AnimatedIndexedStack>
     _selectedIndex = widget.selectedIndex ?? 0;
 
     /// Guards
-    if (_list.length < 2) {
+    if (_list!.length < 2) {
       throw Exception("Atleast 2 children must be provided.");
     }
-    if (_selectedIndex >= _list.length || _selectedIndex < 0) {
+    if (_selectedIndex! >= _list!.length || _selectedIndex! < 0) {
       throw Exception(
-          "Index out of bounds. Selected index must be between index of length of children i.e 0..${_list.length - 1}");
+          "Index out of bounds. Selected index must be between index of length of children i.e 0..${_list!.length - 1}");
     }
 
     ///
 
     if (_prevPage != _selectedIndex) {
-      final currKey = _pageIndexToKeyMap[_selectedIndex];
-      final prevKey = _pageIndexToKeyMap[_prevPage];
+      final currKey = _pageIndexToKeyMap[_selectedIndex!];
+      final prevKey = _pageIndexToKeyMap[_prevPage!];
       Function s = (a, b) {
         if (a.key == currKey) return 2;
         if (b.key == currKey) return -2;
@@ -133,20 +133,20 @@ class _AnimatedIndexedStackState extends State<AnimatedIndexedStack>
       };
       final _sortTime = widget.sortTime ?? SORT_TIME.after;
 
-      if (_sortTime == SORT_TIME.before) _pages.sort(s);
+      if (_sortTime == SORT_TIME.before) _pages.sort(s as int Function(Widget, Widget)?);
 
       // print(_prevPage);
       // print(_selectedIndex);
 
-      _secondaryAnimationControllerList[currKey].reset();
-      if (!_animationControllerList[currKey].isAnimating)
-        _animationControllerList[currKey].reset();
+      _secondaryAnimationControllerList[currKey!]!.reset();
+      if (!_animationControllerList[currKey]!.isAnimating)
+        _animationControllerList[currKey]!.reset();
 
-      _secondaryAnimationControllerList[prevKey].forward().orCancel;
-      _animationControllerList[currKey].forward().orCancel.whenComplete(() {
+      _secondaryAnimationControllerList[prevKey!]!.forward().orCancel;
+      _animationControllerList[currKey]!.forward().orCancel.whenComplete(() {
         if (_sortTime == SORT_TIME.after)
           setState(() {
-            _pages.sort(s);
+            _pages.sort(s as int Function(Widget, Widget)?);
           });
       });
       // print(_pages);
